requirements.txt
asgiref==3.7.2
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Markdown==3.4.4
Pillow==10.0.0
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
psycopg2-binary==2.9.2
dj-rest-auth==4.0.1
django-allauth==0.54.0
dj-rest-auth[with_social]
social-auth-app-django==5.2.0
drf-yasg
django-rest-auth
boto3==1.28.46
django-storages==1.14
django-ckeditor==6.7.0